GraphqlNotebook/docker/src/requirements.txt
aniso8601==7.0.0
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
promise==2.3
pycparser==2.20
PyJWT==2.1.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.0.1
